requirements.txt
streamlit>=1.28.0
boto3>=1.28.0
pillow>=10.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
pypdf2>=3.0.0
python-pptx>=0.6.23
moviepy>=1.0.3
pydub>=0.25.1
requests>=2.31.0
click>=8.1.7
pyyaml>=6.0.1
twilio>=8.10.0
python-dotenv>=1.0.0
jinja2>=3.1.2
markdown>=3.5.1
beautifulsoup4>=4.12.2
lxml>=4.9.3
selenium>=4.15.0
webdriver-manager>=4.0.1
reportlab>=4.0.0

